pyproject.toml
[project]
name = "advent-of-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "more-itertools>=10.6.0",
    "numpy>=2.2.3",
    "progressbar2>=4.5.0",
]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"

[tool.ruff.lint]
extend-select = ['T10', 'E', 'W', 'DJ', 'I', 'UP']
preview = true

[dependency-groups]
dev = ["mypy>=1.15.0", "ruff>=0.9.7"]
